prov/efa: introduce function rxr_pkt_init_data_from_tx_entry()
This patch merged various functions/code snippets that
set up data in a TX packet into a new function:

     rxr_pkt_init_data_from_tx_entry()

and use it to initialize all packets that may contain data,
which include RTM, RTW, DATA, READRSP packets.

Depend on the tx_entry, the function will either copy data to
packet entry, or setup pkt_entry->sendv to point to tx_entry->iov.

The resulted packet entry is ready to be sent by using
rxr_pkt_entry_send().
