+++ prov/efa/src/rxr/rxr_pkt_type_base.c
@@ -32,6 +32,7 @@
  */
 
 #include "efa.h"
+#include "rxr_read.h"
 
 
 /**
@@ -62,4 +63,90 @@ uint32_t *rxr_pkt_connid_ptr(struct rxr_pkt_entry *pkt_entry)
 	}
 
 	return NULL;
+}
+
+/**
+ * @brief set up data in a packet entry using tx_entry information, such that the packet is ready to be sent.
+ *        Depend on the tx_entry, this function can either copy data to packet entry, or point
+ *        pkt_entry->iov to tx_entry->iov.
+ *        It requires the packet header to be set.
+ *
+ * @param[in]		ep		end point.
+ * @param[in,out]	pkt_entry	packet entry. Header must have been set when the function is called
+ * @param[in]		hdr_size	packet header size.
+ * @param[in]		tx_entry	This function will use iov, iov_count and desc of tx_entry
+ * @param[in]		data_offset	offset of the data to be set up. In reference to tx_entry->total_len.
+ * @param[in]		data_size	length of the data to be set up. In reference to tx_entry->total_len.
+ * @return		no return
+ */
+void rxr_pkt_init_data_from_tx_entry(struct rxr_ep *ep,
+				     struct rxr_pkt_entry *pkt_entry,
+				     size_t hdr_size,
+				     struct rxr_tx_entry *tx_entry,
+				     size_t data_offset,
+				     size_t data_size)
+{
+	int tx_iov_index;
+	char *data;
+	size_t tx_iov_offset, copied;
+	struct efa_mr *desc;
+
+	assert(hdr_size > 0);
+
+	pkt_entry->x_entry = tx_entry;
+	/* pkt_sendv_pool's size equal efa_tx_pkt_pool size +
+	 * shm_tx_pkt_pool size. As long as we have a pkt_entry,
+	 * pkt_entry->send should be allocated successfully
+	 */
+	pkt_entry->send = ofi_buf_alloc(ep->pkt_sendv_pool);
+	if (!pkt_entry->send)
+		FI_WARN(&rxr_prov, FI_LOG_EP_CTRL, "allocate pkt_entry->send failed\n");
+	assert(pkt_entry->send);
+
+	if (data_size == 0) {
+		pkt_entry->send->iov_count = 0;
+		pkt_entry->pkt_size = hdr_size;
+		return;
+	}
+
+	rxr_locate_iov_pos(tx_entry->iov, tx_entry->iov_count, data_offset,
+			   &tx_iov_index, &tx_iov_offset);
+	desc = tx_entry->desc[0];
+	assert(tx_iov_index < tx_entry->iov_count);
+	assert(tx_iov_offset < tx_entry->iov[tx_iov_index].iov_len);
+
+	/*
+	 * Copy can be avoid if the following 2 conditions are true:
+	 * 1. user provided memory descriptor, or message is sent via shm provider
+	 *    (which does not require a memory descriptor)
+	 * 2. data to be send is in 1 iov, because device only support 2 iov, and we use
+	 *    1st iov for header.
+	 */
+	if ((!pkt_entry->mr || tx_entry->desc[tx_iov_index]) &&
+	    (tx_iov_offset + data_size <= tx_entry->iov[tx_iov_index].iov_len)) {
+
+		assert(ep->core_iov_limit >= 2);
+		pkt_entry->send->iov[0].iov_base = pkt_entry->pkt;
+		pkt_entry->send->iov[0].iov_len = hdr_size;
+		pkt_entry->send->desc[0] = pkt_entry->mr ? fi_mr_desc(pkt_entry->mr) : NULL;
+
+		pkt_entry->send->iov[1].iov_base = (char *)tx_entry->iov[tx_iov_index].iov_base + tx_iov_offset;
+		pkt_entry->send->iov[1].iov_len = data_size;
+		pkt_entry->send->desc[1] = tx_entry->desc[tx_iov_index];
+		pkt_entry->send->iov_count = 2;
+		pkt_entry->pkt_size = hdr_size + data_size;
+		return;
+	}
+
+	data = pkt_entry->pkt + hdr_size;
+	copied = ofi_copy_from_hmem_iov(data,
+					data_size,
+					desc ? desc->peer.iface : FI_HMEM_SYSTEM,
+					desc ? desc->peer.device.reserved : 0,
+					tx_entry->iov,
+					tx_entry->iov_count,
+					data_offset);
+	assert(copied == data_size);
+	pkt_entry->send->iov_count = 0;
+	pkt_entry->pkt_size = hdr_size + copied;
 }
